Make code examples consistent with map examples
By default, many of the Playground maps use the 'bar' style
search control, but the code examples associated with the maps
do not specify the style. This ends up rendering a map with a
button-style control, which does not match the example maps.

This commit updates the documentation to specify `style: 'bar'`
whenever the associated example map is using a bar.