feat: add handling of lat, lon coordinate queries (#368)
Quick, *pure js* implementation to handle coordinate queries (#147). If
a user enters valid coordinates, they get a single result with that
point (as in Google Maps and other services). It's a useful feature for
power users.

![image](https://github.com/smeijer/leaflet-geosearch/assets/8945883/775ac7a2-3802-44e8-a3d5-c59d91d579e6)

Any other queries are sent to the Provider. Happy to see this in
TypeScript and otherwise improved, it's probably not ready for merging.

---------

Co-authored-by: Stephan Meijer <stephan.meijer@gmail.com>